chore: upgrace demo to Next 15, upgrade solana packages, resolve demo playwright issues
