+++ apps/demo/next.config.mjs
@@ -2,9 +2,6 @@ import { withSentryConfig } from '@sentry/nextjs'
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  eslint: {
-    ignoreDuringBuilds: true
-  },
   distDir: 'out',
   cleanDistDir: true,
   images: {
