+++ apps/demo/package.json
@@ -3,11 +3,11 @@
   "version": "1.1.7",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
-    "start": "next start --port 3003",
+    "start": "next start",
     "playwright:start": "pnpm start",
-    "playwright:test": "HEADLESS=true playwright test",
+    "playwright:test": "playwright test",
     "playwright:debug": "pnpm playwright:test --debug"
   },
   "dependencies": {
@@ -39,6 +39,7 @@
     "@sentry/core": "8.49.0",
     "@sentry/nextjs": "8.49.0",
     "@solana/wallet-adapter-wallets": "0.19.32",
+    "@solana/wallet-adapter-base": "0.9.26",
     "@solana/web3.js": "1.98.0",
     "@tanstack/react-query": "5.56.2",
     "@wagmi/core": "2.16.7",
@@ -47,7 +48,7 @@
     "classnames": "2.5.1",
     "clsx": "2.1.1",
     "ethers": "6.13.5",
-    "next": "14.2.28",
+    "next": "15.3.1",
     "next-themes": "0.4.3",
     "react": "19.0.0",
     "react-dom": "19.0.0",
@@ -62,7 +63,7 @@
     "@types/node": "22.13.4",
     "@types/react": "19.0.0",
     "@types/react-dom": "19.0.0",
-    "@playwright/test": "1.48.2",
+    "@playwright/test": "1.52.0",
     "eslint": "8.56.0",
     "eslint-config-next": "14.1.0",
     "postcss": "8.5.3",
