+++ apps/demo/tests/constants.ts
@@ -1,1 +1,1 @@
-export const BASE_URL = process.env['BASE_URL']!
+export const BASE_URL = process.env['BASE_URL'] || 'http://localhost:3000/'
