+++ packages/adapters/solana/package.json
@@ -52,7 +52,7 @@
     "@reown/appkit-polyfills": "workspace:*",
     "@reown/appkit-utils": "workspace:*",
     "@reown/appkit-wallet": "workspace:*",
-    "@solana/wallet-adapter-base": "0.9.24",
+    "@solana/wallet-adapter-base": "0.9.26",
     "@solana/wallet-standard-features": "1.3.0",
     "@solana/wallet-standard-util": "1.1.2",
     "@solana/web3.js": "1.98.0",
