+++ packages/appkit-utils/package.json
@@ -56,7 +56,7 @@
   },
   "devDependencies": {
     "@coinbase/wallet-sdk": "4.3.0",
-    "@solana/wallet-adapter-base": "0.9.24",
+    "@solana/wallet-adapter-base": "0.9.26",
     "@solana/web3.js": "1.98.0",
     "@vitest/coverage-v8": "2.1.9",
     "vitest": "2.1.9"
